+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
+- Add automatic reloading of .po and .mo files in development mode
+
 ## 2.0.0
+
 - change how model attributes are looked up (class first, then sti root)
 - drop support for old rubies
 
 ## v1.13.0
+
 - Use subclasses instead of direct_descendants on rails 7 and above
 
 ## v1.12.0